v2017.02.11
Initial publication of 2 extensions.

- Add Class Monitor: same as add-sample-class which can be found on
iTop's official documentation pages; but this version includes model
type 'Monitor', which is otherwise bugged.

- Fix Class IP Phone: this completely redefines the existing IPPhone
class. Rather than to view the IP Phone as a variant of a Phone, it's
now seen as a variant of a Network Device. This allows to connect to
other devices (since IP Phones are often also a small switch) and it
adds three fields: management IP, internal and external number.